Update for RT 4.4.
Much of this is just rote work to follow the new Mail Plugin API.  The
biggest change is that we hook in much earlier in the process, so we don't
have a proper ticket object to work on.  Instead we work directly on the
MIME Message object, including modifying the subject to tie the message to
a ticket.